feat(deps): add support for Nix package manager
This serves as an alternative to Homebrew. It should be much more stable
and cause less headaches over time for automated builds.

There should be no change to the end user experience of using the build
script, as it should still work with and use Homebrew by default.

Additionally, Nix provides older Apple SDK's, so builds run against
macOS 11.x SDKs via Nix. This should in theory allow the resulting
Emacs.app builds should be compatible with older macOS versions. Exactly
how well that holds up in practice remains to be seen.

Nix does support customizing the Apple SDK version, but I have not yet
figured out how to do so via a flake, and it's not a priority at the
moment, as the default v11 SDK is sufficient.