wip(libs): attempt to fix code signing issue on Intel-based macs
This is a horrible hack attempt to fix a application signing
verification bug with the x86_64 builds. It seems that most of the Intel
based libraries from Nix uses `LC_VERSION_MIN_MACOSX`, most specify both
`version` and `sdk` values. However some don't specify a `sdk` value,
and when checked with `codeside -vv -d <path/to/file>`, it complains
with a warning:

    Library validation warning=OS X SDK version before 10.9 does not support Library Validation

I suspect this is why the Intel builds fail signing verification. This
is an attempt to fix that. It uses `vtool` to force set the newer
`LC_BUILD_VERSION` load command's `minos` and `sdk` values to that of
the old load command's `version` value. But only if the old load command
is present and missing an `sdk` value.